pyproject.toml
[project]
name = "promptstrap"
version = "0.1.0"
description = ""
authors = [
    {name = "bigfatfrodo",email = "133560036+bigfatfrodo@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "ollama (>=0.5.1,<0.6.0)",
    "langchain-openai (>=0.3.23,<0.4.0)",
]

[tool.poetry]
packages = [{include = "promptstrap", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
